requirements.txt
# Core dependencies
flask==3.1.1
werkzeug==3.0.1
pillow==10.0.1

# AI and ML dependencies
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0

# TTS and audio processing
tortoise-tts>=2.7.0
librosa>=0.10.0
soundfile>=0.12.0

# Computer vision
opencv-python>=4.8.0
opencv-python-headless>=4.8.0

# Utilities
requests>=2.31.0
psutil>=5.9.0
tqdm>=4.65.0

# AI/ML frameworks
transformers==4.31.0
accelerate>=0.20.0
huggingface-hub>=0.16.0
safetensors>=0.3.0

# Additional dependencies for Wav2Lip
scikit-image>=0.19.0
scikit-learn>=1.3.0
basicsr>=1.4.0
facexlib>=0.3.0
gfpgan>=1.3.0
realesrgan>=0.3.0 